+++ documents/assistantResponse_class.md
@@ -1,14 +1,20 @@
-# `assistantResponse`
-I did refactoring to collect all the logics about `response` to `assistantResponse` class.
+# assistantResponse Class
 
-You can access a instance of this class with `this.assistantResponse` of `MMM-AssistantMk2`
+```
+/library/
+  response.class.js
+  (response.js) <- default
+/library/responseUI
+  response.js samples per UI
+```
 
-I tried to keep many legacy codes from before-refactoring, but many parts would be missing. especially about icon and chime.
+User need to copy one of `/library/responseUI/respose.samples.js` then paste it into `/library/` as `response.js`
+By example;
+- response.fullscreen.js
+- response.800x600.js
+- response.onlyicon.js
+- response.simple.js
 
-Major pulic methods will be these. (Usable in outside of class)
 
-- AssistantResponse.tunnel(tunnel_payload)
-- AssistantResponse.playChime(sound)
-- AssistantResponse.status(statusText)
-
-For other methods, I hope you don't need. Javascript doesn't support encapsulation. If was possible, I would have made all of methods as `private`.
+`response.js`(**AssistantResponse**) should be child-class of `response.class.js`(**AssistantResponseClass**).
+So you just override members what UI needs.
