+++ library/response.js
@@ -1,0 +1,70 @@
+class AssistantResponse extends AssistantResponseClass{
+  constructor (responseConfig, callbacks) {
+    super(responseConfig, callbacks)
+  }
+
+
+//You can implement these methods by your needs.
+/*
+  tunnel (payload) {
+    super(payload)
+  }
+
+  setSecret (secretMode) {
+    super(secretMode)
+  }
+
+  playChime (sound) {
+    super(sound)
+  }
+
+  status (status) {
+    super(status)
+  }
+
+  prepare () {
+    super()
+  }
+
+  getDom () {
+    return super()
+  }
+
+  showError (text) {
+    super(text)
+  }
+
+  showTranscription (text, className = "transcription") {
+    super(text, className)
+  }
+
+  end () {
+    super()
+  }
+
+  start (response) {
+    super(response)
+  }
+
+  stopResponse (callback = ()=>{}) {
+    super(callback)
+  }
+
+  postProcess (response, callback_done=()=>{}, callback_none=()=>{}) {
+    super(response, callback_done, callback_none)
+  }
+
+  playAudioOutput (response) {
+    return super(response)
+  }
+
+  showScreenOutput (response) {
+    return super(response)
+  }
+
+  makeUrl (uri) {
+    return super(uri)
+  }
+
+*/
+}
