+++ screen_output.css
@@ -11,6 +11,7 @@ html,body{background:transparent;margin:0}
 .inactive{opacity:.4}
 #carousel-container{left:0;overflow-x:initial;transition:300ms ease-in-out;transform:translateZ(0)}
 /* Add here your overriding CSS */
-body {zoom:55%;}
+body {zoom:50%;}
 .assistant-bar-container {display:none;}
 #assistant-shadow {display:none;}
+.scroll-bar {display:none;}
